+++ src/main/java/affableBean/repository/OrderedProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import affableBean.domain.OrderedProduct;
 
 @Repository
-public interface OrderedProductRepository extends JpaRepository<OrderedProduct, Long> {
+public interface OrderedProductRepository extends JpaRepository<OrderedProduct, Integer> {
 	
 }
